requirements.txt
pandas==2.2.3
biopython==1.85
